btorskel: Delay adding new assertions.
When going through the ids table and checking the values of the literals
we assume that if the value is unknown the corresponding expression
can't be a constraint.
For option combination sort-exp=0 and simplify-constraints=0 this is not
true anymore since btor_assert_exp may trigger new assertions to be
added that are already in the ids table, but don't have a value yet.
This commit delays adding new assertions.