Fix simplified assumption handling.
Previously, we always stored the simplified versions of assumptions in
the assumptions hash table. However, this can result in assumptions
being removed from the assumptions table after a check-sat call, which
breaks querying failed assumptions.
As a consquence, we now keep the original assumptions and don't update
the assumptions table.